+++ packages/eas-cli/src/worker/fragments/WorkerDeployment.ts
@@ -1,0 +1,11 @@
+import gql from 'graphql-tag';
+
+export const WorkerDeploymentFragmentNode = gql`
+  fragment WorkerDeploymentFragment on WorkerDeployment {
+    id
+    url
+    deploymentIdentifier
+    deploymentDomain
+    createdAt
+  }
+`;